Refactor Addon sidebar into app/sidebar to aid reuse
