--- src/pages/app/sidebar/ProjectList.container.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-  import Projects from "./ProjectList.svelte";
-  import { projects } from "../../../manager/projects.js";
-  import { newProject, editProject } from "../../../manager/layout.js";
-</script>
-
-<Projects {projects} {newProject} {editProject} />
