+++ src/pages/app/sidebar/projects/stories/Project.stories.svelte
@@ -1,4 +1,4 @@
-<script context="module">
+<script lang="ts" context="module">
   import { Template, Story } from "@storybook/addon-svelte-csf";
   import { action } from "@storybook/addon-actions";
   import Project from "../Project.svelte";
