+++ src/pages/app/sidebar/stories/DocumentFilters.stories.svelte
@@ -1,28 +1,27 @@
-<script context="module">
-  import { Story } from "@storybook/addon-svelte-csf";
-  import DocumentFiltersComponent from "../DocumentFilters.svelte";
+<script lang="ts" context="module">
+  import { Template, Story } from "@storybook/addon-svelte-csf";
+  import DocumentFilters from "../DocumentFilters.svelte";
 
-  import * as mock from "../../stories/mock.js";
+  import userFixture from "../../fixtures/me.json";
 
   export const meta = {
     title: "App / Sidebar / Document Filters",
-    component: DocumentFiltersComponent,
+    component: DocumentFilters,
     parameters: { layout: "centered" },
+  };
+
+  const args = {
+    user: null,
   };
 </script>
 
+<Template let:args>
+  <DocumentFilters {...args} />
+</Template>
+
+<Story name="Without User" {args} />
 <Story
-  name="Document Filters"
-  parameters={{
-    msw: {
-      handlers: [
-        mock.users.data,
-        mock.users.me,
-        mock.organizations.data,
-        mock.projects.data,
-      ],
-    },
-  }}
->
-  <DocumentFiltersComponent />
-</Story>
+  name="With User"
+  args={{ user: { ...userFixture, organization: null } }}
+/>
+<Story name="With Org User" args={{ user: userFixture }} />
